+++ HSTracker/Logging/Enums/PlacementMode.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  HSTracker
+//
+//  Created by Francisco Moraes on 4/4/25.
+//  Copyright © 2025 Benjamin Michotte. All rights reserved.
+//
+
+import Foundation
+
+enum PlacementMode: Int {
+    case bottom, right, left, top
+}